// Converge/ClientApp/src/types/PresenceAvailability.ts
// Copyright (c) Microsoft Corporation.
// Licensed under the MIT License.

enum PresenceAvailability {
  Available = "Available",
  AvailableIdle = "AvailableIdle",
  Away = "Away",
  BeRightBack = "BeRightBack",
  Busy = "Busy",
  BusyIdle = "BusyIdle",
  DoNotDisturb = "DoNotDisturb",
  Offline = "Offline",
  PresenceUnknown = "PresenceUnknown",
  OutOfOffice = "OutOfOffice"
}

export default PresenceAvailability;